Don't choke on directories

+++ core/src/main/java/com/github/swissquote/carnotzet/core/maven/ResourcesManager.java
@@ -63,7 +63,11 @@ public class ResourcesManager {
 			for (Path source : (Iterable<Path>) walk::iterator) {
 				Path destination = Paths.get(destinationDirectoryLocation, source.toString()
 						.substring(sourceDirectoryLocation.length()));
-				Files.copy(source, destination, StandardCopyOption.REPLACE_EXISTING);
+
+				// Don't try to overwrite directories that already exist
+				if (!(Files.isDirectory(source) && Files.exists(destination))) {
+					Files.copy(source, destination, StandardCopyOption.REPLACE_EXISTING);
+				}
 			}
 		}
 	}
