+++ core/src/test/java/com/github/swissquote/carnotzet/core/maven/ResourcesManagerTest.java
@@ -34,7 +34,11 @@ public class ResourcesManagerTest {
 			for (Path source : (Iterable<Path>) walk::iterator) {
 				Path destination = Paths.get(destinationDirectoryLocation, source.toString()
 						.substring(sourceDirectoryLocation.length()));
-				Files.copy(source, destination, StandardCopyOption.REPLACE_EXISTING);
+
+				// Don't try to overwrite directories that already exist
+				if (!(Files.isDirectory(source) && Files.exists(destination))) {
+					Files.copy(source, destination, StandardCopyOption.REPLACE_EXISTING);
+				}
 			}
 		}
 	}
